first/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==5.2.3
django-cors-headers==4.9.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Markdown==3.8.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
redis==7.0.1
sqlparse==0.5.3
typing_extensions==4.14.0
